Updated based on PR #43 feedback
